expense-tracker/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
PyJWT==2.8.0
werkzeug==3.0.3
pytest==8.2.2
pytest-flask==1.3.0
